Update to fix issues with running on Linux
- Updated to fix error with ServiceProcess when running on Linux
- Added error message if Windows Service commands are attempted on Linux

+++ applications/Agents/MTConnect-Agent-Http-Persistent/Program.cs
@@ -60,7 +60,7 @@ namespace MTConnect.Applications
         {
             PrintHeader();
 
-            string command = "debug";
+            string command = "run";
             string configFile = null;
             int port = 0;
 
@@ -115,8 +115,14 @@ namespace MTConnect.Applications
                     while (true) System.Threading.Thread.Sleep(100); // Block (exit console by 'Ctrl + C')
 
                 case "run-service":
+
                     _verboseLogging = true;
-                    ServiceBase.Run(service);
+                    if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
+                    {
+                        ServiceBase.Run(service);
+                    }
+                    else _applicationLogger.Info($"'Run-Service' Command is not supported on this Operating System");
+
                     break;
 
                 case "debug":
@@ -125,32 +131,67 @@ namespace MTConnect.Applications
                     while (true) System.Threading.Thread.Sleep(100); // Block (exit console by 'Ctrl + C')
 
                 case "install":
-                    service.StopService();
-                    service.RemoveService();
-                    service.InstallService(configFile);
+
+                    if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
+                    {
+                        service.StopService();
+                        service.RemoveService();
+                        service.InstallService(configFile);
+                    }
+                    else _applicationLogger.Info($"'Install' Command is not supported on this Operating System");
+
                     break;
 
                 case "install-start":
-                    service.StopService();
-                    service.RemoveService();
-                    service.InstallService(configFile);
-                    service.StartService();
+
+                    if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
+                    {
+                        service.StopService();
+                        service.RemoveService();
+                        service.InstallService(configFile);
+                        service.StartService();
+                    }
+                    else _applicationLogger.Info($"'Install-Start' Command is not supported on this Operating System");
+
                     break;
 
                 case "remove":
-                    service.StopService();
-                    service.RemoveService();
+
+                    if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
+                    {
+                        service.StopService();
+                        service.RemoveService();
+                    }
+                    else _applicationLogger.Info($"'Remove' Command is not supported on this Operating System");
+
                     break;
 
                 case "start":
-                    service.StartService();
+
+                    if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
+                    {
+                        service.StartService();
+                    }
+                    else _applicationLogger.Info($"'Start' Command is not supported on this Operating System");
+
                     break;
 
                 case "stop":
-                    service.StopService();
+
+                    if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
+                    {
+                        service.StopService();
+                    }
+                    else _applicationLogger.Info($"'Stop' Command is not supported on this Operating System");
+
                     break;
 
-                case "help": 
+                case "help":
+                    PrintHelp();
+                    break;
+
+                default:
+                    _applicationLogger.Info($"'{command}' : Command not recognized : See help for more information");
                     PrintHelp();
                     break;
             }
